Update SlimCD.php
This change adds support for TLS 1.2, which SLIM CD will be using by June 2018.  Note that I also commented out the disabling of the VERIFYPEER checking so that SSL connections will be verified.  This will force developers to install the CA certificate for GoDaddy.  if you cannot install that certificate in your environment, you can uncomment the VERIFYPEER disabling, but this is NOT recommended!